saving 1st attempt

+++ predict_module/supervised_finetune.py
@@ -69,7 +69,7 @@ def supervised_finetune(args):
     # tokenizer.padding_side = "left"  # Allow batched inference
 
     data = load_dataset("json", data_files=DATA_PATH)
-    val_set_size = VAL_PCT * len(data)
+    val_set_size = int(VAL_PCT * len(data))
 
     now_max_steps = max(
         (len(data["train"]) - val_set_size) // BATCH_SIZE * EPOCHS, EPOCHS)
